Изменил расписание удаления старых записей в базе данных

+++ port_status.py
@@ -130,10 +130,12 @@ def main():
     print('Необработанные: {}'.format(', '.join(dslam_bad)))
     print('---------\n')
             
-    # Удаление старых записей
-    options = {'table_name': 'data_dsl',
-               'str1': 'CAST(datetime AS DATE) < DATE_ADD(CURRENT_DATE(), INTERVAL -{} DAY)'.format(Settings.days)}
-    SQL.delete_table(**options)
+    # Удаление старых записей (раз в день в промежутке между 0 и 2 часами)
+    hour_now = datetime.datetime.now().hour
+    if (hour_now >= 0) and (hour_now < 2):
+        options = {'table_name': 'data_dsl',
+                   'str1': 'CAST(datetime AS DATE) <= DATE_ADD(CURRENT_DATE(), INTERVAL -{} DAY)'.format(Settings.days)}
+        SQL.delete_table(**options)
